Prevent erasure of the whole directory
git-svn-id: file:///home/roberto/Git/svn@5525 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: fd63ebda2bf15f7dfa1b3a055d7ad0ada03db8c9

+++ trunk/xmipp/applications/scripts/protocols/protocol_kerdensom.py
@@ -141,6 +141,8 @@ class kerdensom_class:
                 
         # Delete working directory if it exists, make a new one
         if (_DoDeleteWorkingDir): 
+            if (self.WorkingDir==""):
+               raise RuntimeError,"No working directory given"
             if os.path.exists(self.WorkingDir):
                 shutil.rmtree(self.WorkingDir)
         if not os.path.exists(self.WorkingDir):
