+++ trunk/xmipp/applications/scripts/protocols/protocol_ml3d.py
@@ -272,6 +272,8 @@ class ML3D_class:
         if (RestartIter < 1):
             # Delete working directory if it exists, make a new one
             if (DoDeleteWorkingDir): 
+                if (self.WorkingDir==""):
+                    raise RuntimeError,"No working directory given"
                 if os.path.exists(self.WorkingDir):
                     shutil.rmtree(self.WorkingDir)
             if not os.path.exists(self.WorkingDir):
