+++ trunk/xmipp/applications/scripts/protocols/protocol_rct.py
@@ -163,6 +163,8 @@ class RCT_class:
                 
         # Delete working directory if it exists, make a new one, and go there
         if (DoDeleteWorkingDir): 
+            if (self.WorkingDir==""):
+                raise RuntimeError,"No working directory given"
             if os.path.exists(self.WorkingDir):
                 shutil.rmtree(self.WorkingDir)
         if not os.path.exists(self.WorkingDir):
